feat(read config from config.py):
