+++ api/main.py
@@ -1,8 +1,10 @@
 from flask import Flask
 from flask_restful import Api
 from resources import ems
+from resources import app
 
-app = Flask(__name__)
+# app = Flask(__name__)
+# app.config.from_pyfile('config.py')
 api = Api(app)
 # Hems
 api.add_resource(ems.Hems,
@@ -11,7 +13,6 @@ api.add_resource(ems.Hems,
 # Bems
 api.add_resource(ems.Bems,
                  "/bems/<string:name>")
-
 
 if __name__ == "__main__":
     app.run(debug='no')
