+++ api/resources/__init__.py
@@ -1,0 +1,3 @@
+from flask import Flask
+app = Flask(__name__)
+app.config.from_pyfile('config.py')
