+++ api/resources/config.py
@@ -1,0 +1,3 @@
+HEMS_ACCEPT = ['aggregator_distribution']
+BEMS_ACCEPT = ['appliances']
+
